moved compliance tests to ComplianceTest class; lightened GroupingTest.php
